requirements.txt
pygame
numpy
matplotlib
pytorch
gym[box2d]
gymnasium[mujoco]